Extracted localStorageDataConnection and searchLocalStorageDataConnections
